requirements.txt
asgiref==3.8.1
astroid==3.3.9
babel==2.17.0
colorama==0.4.6
dill==0.3.9
Django==4.2.20
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
isort==6.0.1
mccabe==0.7.0
platformdirs==4.3.6
psycopg==3.2.6
psycopg-binary==3.2.6
py-moneyed==3.0
PyJWT==2.9.0
pylint==3.3.5
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
python-dotenv==1.1.0
setuptools==78.1.0
sqlparse==0.5.3
tomlkit==0.13.2
typing_extensions==4.13.0
tzdata==2025.1
